Reorg everything into private modules
